requirements.txt
absl-py==2.1.0
asgiref==3.8.1
astunparse==1.6.3
certifi==2025.1.31
charset-normalizer==3.4.1
cmdstanpy==1.2.5
contourpy==1.3.1
crispy-bootstrap5==0.6
cycler==0.12.1
diff-match-patch==20241021
Django==5.0.12
django-crispy-forms==1.14.0
django-import-export==4.3.7
djangorestframework==3.15.2
et_xmlfile==2.0.0
flatbuffers==25.2.10
fonttools==4.56.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.71.0
h5py==3.13.0
holidays==0.68
idna==3.10
importlib_resources==6.5.2
joblib==1.4.2
keras==3.9.0
kiwisolver==1.4.8
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
ml_dtypes==0.5.1
namex==0.0.8
numpy==2.1.3
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.14.1
packaging==24.2
pandas==2.2.3
pillow==11.1.0
prophet==1.1.6
protobuf==5.29.3
psycopg==3.2.5
psycopg-binary==3.2.5
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
requests==2.32.3
rich==13.9.4
scikit-learn==1.6.1
scipy==1.15.2
setuptools==75.8.0
six==1.17.0
sqlparse==0.5.3
stanio==0.5.1
tablib==3.8.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==2.5.0
threadpoolctl==3.5.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
xlrd==2.0.1
